Finish breaking up the `test_wikimedia.py` file
