+++ tests/apis/wikimedia/test_language_methods.py
@@ -1,3 +1,4 @@
+from flickypedia.apis import WikimediaApi
 from flickypedia.apis.wikimedia.language_methods import order_language_list
 
 
@@ -30,3 +31,21 @@ def test_order_language_list() -> None:
         {"id": "sco", "label": "Scots", "match_text": "escocès"},
         {"id": "myv", "label": "эрзянь", "match_text": "esiya — èdè esiya"},
     ]
+
+
+def test_find_matching_languages(wikimedia_api: WikimediaApi) -> None:
+    actual = wikimedia_api.find_matching_languages(query="des")
+    expected = [
+        {"id": "da", "label": "dansk", "match_text": "dens"},
+        {
+            "id": "ase",
+            "label": "American sign language",
+            "match_text": "des — langue des signes américaine",
+        },
+    ]
+
+    assert actual == expected
+
+
+def test_handles_no_matching_languages(wikimedia_api: WikimediaApi) -> None:
+    assert wikimedia_api.find_matching_languages(query="ymr") == []
